prime-sms-hub/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ>=0.12.0
python-decouple==3.8
requests==2.31.0
firebase-admin==6.2.0
paystackapi>=1.7.3
sendgrid==6.10.0
livereload>=2.6.4
watchdog==3.0.0
Pillow==10.1.0
channels==4.1.0
daphne==4.0.0
